renaming task editor related files
