+++ qtodotxt/ui/dialogs/taskeditor_lineedit.py
